third_party/blink/web_tests/http/tests/devtools/file-system-project-expected.txt

Tests file system project.


Running: testFileSystems
Adding first file system.
Adding second file system.
Adding files to file systems.

UISourceCode: file:///var/www/localhost/foo.js
UISourceCode is content script: false
Highlighter type: text/javascript
UISourceCode content: <foo content>

UISourceCode: file:///var/www/bar.js
UISourceCode is content script: false
Highlighter type: text/javascript
UISourceCode content: <bark content>

UISourceCode: file:///foo/bar/baz.js
UISourceCode is content script: false
Highlighter type: text/javascript
UISourceCode content: <bazzz content>

Dumping uiSourceCode location link texts:
 - foo.js:6
 - bar.js:6
 - baz.js:6
After revision added:
UISourceCode: file:///var/www/localhost/foo.js
UISourceCode is content script: false
Highlighter type: text/javascript
UISourceCode content: <Modified UISourceCode content>

Running: testDefaultExcludes

-- Default excludes --
Dumping uiSourceCodes origin URLs:
  - file:///var/www/html/foo.js
  - file:///var/www/bar.js
  - file:///var/www/html2/foo.js

Running: testExcludesSettings

-- Excluded /html/ --
Dumping uiSourceCodes origin URLs:
  - file:///var/www2/bar.js
  - file:///var/www2/html2/foo.js

Running: testExcludesViaDelegate

-- Excluded /html2/ --
Dumping uiSourceCodes origin URLs:
  - file:///var/www3/html/foo.js
  - file:///var/www3/bar.js

Running: testFileAddedExternally
-- Original tree --
Dumping uiSourceCodes origin URLs:
  - file:///var/www4/html/foo.js
-- File added externally --
Dumping uiSourceCodes origin URLs:
  - file:///var/www4/html/foo.js
  - file:///var/www4/html/bar.js

Running: testGitFolders
project_1/
project_2/

Running: testUISourceCodeMetadata
Initial metadata
    content size: 6
    modification time: 1970-01-01T00:16:40.000Z
Changed metadata
    content size: 15
    modification time: 1970-01-01T00:16:41.000Z

Running: testFileRename
URL before rename: file:///var/www3/test/hello.js
URL after rename: file:///var/www3/test/goodbye.js
